+++ common/src/main/java/ru/sbrf/common/PaymentRequest.java
@@ -1,7 +1,6 @@
 package ru.sbrf.common;
 
 import lombok.*;
-
 import java.util.Date;
 
 @AllArgsConstructor
@@ -13,7 +12,7 @@ public class PaymentRequest {
     private Date date;
     private String numberAccount;
     private String numberPhone;
-    private long summa;
+    private int summa;
     private Currency currency;
 }
